Move sendCommentEmail to email module
