+++ src/routes.js
@@ -6,7 +6,6 @@ var http = require('http');
 var url = require('url');
 var path = require('path');
 var openid = require('openid');
-var _ = require('underscore');
 
 var Firebase = require('firebase');
 var FirebaseTokenGenerator = require("firebase-token-generator");
@@ -54,138 +53,19 @@ router.get('/', function (req, res) {
 });
 
 router.get('/email/sendCommentEmail', function (req, res) {
-    var requestid = req.query.requestid;
+    var requestId = req.query.requestid;
     var commenter = req.query.commenter;
     var creator = req.query.creator;
-    var gametitle = req.query.gametitle;
+    var gameTitle = req.query.gametitle;
     var system = req.query.system;
 
-    var titlemessage = 'Someone commented on your request';
-    var grammarfocus = 'your';
-    var interestfocus = 'you are interested in'
-    var subject = 'Someone commented on your gaming request';
-
-    //Send email to the owner of the request
-    var requesturl = origin + '/#/request/' + requestid;
-
-    //Gonna need to do some querying - Get the creator, then for each subscriber, get their user info, and send an email to them if they are subscribed
-    var creatorRef = new Firebase(config.firebase.url + 'users');
-
-    creatorRef
-    .orderByChild('username')
-    .equalTo(creator)
-    .once('value', function (creatorSnapshot) {
-        var creatorUserObj = creatorSnapshot.val();
-
-        if (creatorUserObj !== null && creatorUserObj !== undefined) {
-
-            var creatorEmail = null;
-            var creatorUser = null;
-
-            var creatorUserArray = _.map(creatorUserObj, function (userItem, userKey) {
-                return userItem;
-            });
-
-            if (creatorUserArray !== null && creatorUserArray !== undefined && creatorUserArray.length > 0) {
-                creatorUser = creatorUserArray[0];
-                creatorEmail = creatorUserArray[0].email;
-            }
-
-            //Send an email to the creator if (a) their email is verified & (b) they are not the commenter
-            if (
-                creatorEmail !== null &&
-                creatorEmail !== undefined &&
-                creatorUser !== null &&
-                creatorUser !== undefined &&
-                creatorUser.emailverified !== null &&
-                creatorUser.emailverified !== undefined &&
-                creatorUser.emailverified == true &&
-                creator.trim() !== commenter.trim()) {
-
-                //send email to creator
-                var params = {
-	                addressee: creator,
-	                titlemessage: titlemessage,
-	                grammarfocus: grammarfocus,
-	                interestfocus: '',  //interest focus is empty for the creator
-	                commenter: commenter,
-	                gametitle: gametitle,
-	                system: system,
-	                requesturl: requesturl
-                };
-
-                email.send(creatorUser.email, subject, 'commentRequestEmail.html', params, function (error, info) {
-                    if (error) {
-                        return console.log(error);
-                    }
-                    console.log('Message sent: ' + info.response);
-                });
-            }
-
-            //Now we are going to make a call for each subscribed user, and send them an email one after the other
-            var requestSubscribersRef = new Firebase(config.firebase.url + 'requests/' + requestid + '/subscribers');
-
-            requestSubscribersRef.once('value', function (subsSnapshot) {
-                var subscriberIds = subsSnapshot.val();
-
-                if (subscriberIds !== null && subscriberIds !== undefined) {
-                    var subscriberIdsArray = _.map(subscriberIds, function (sub, key) {
-                        return sub.uid;
-                    });
-
-                    if (subscriberIdsArray !== null && subscriberIdsArray !== undefined && subscriberIdsArray.length > 0) {
-                        //Get the user for each uid and send them an email - all except the commenter
-                        _.each(subscriberIdsArray, function (subId) {
-
-                            var userRef = new Firebase(config.firebase.url + 'users/' + subId);
-                            userRef.once('value', function (userSnapshot) {
-                                var user = userSnapshot.val();
-
-                                if (
-                                    user !== null &&
-                                    user !== undefined &&
-                                    user.username !== null &&
-                                    user.username !== undefined &&
-                                    user.email !== null &&
-                                    user.email !== undefined &&
-                                    user.emailverified !== null &&
-                                    user.emailverified !== undefined &&
-                                    user.emailverified == true &&
-                                    user.username.trim() !== commenter.trim()) {
-
-                                    var params = {
-						                addressee: user.username,
-						                titlemessage: 'Someone commented on a request you are interested in',
-						                grammarfocus: 'a',
-						                interestfocus: 'you are interested in',
-						                commenter: commenter,
-						                gametitle: gametitle,
-						                system: system,
-						                requesturl: requesturl
-					                };
-
-					                email.send(
-					                	user.email,
-					                	'Someone commented on a gaming request you are interested in',
-					                	'commentRequestEmail.html',
-					                	params,
-					                	function (error, info) {
-						                    if (error) {
-						                        return console.log(error);
-						                    }
-						                    console.log('Message sent: ' + info.response);
-					                });
-                                }
-                            });
-                        });
-                    }
-                }
-            });
-        }
-    });
-
-    res.status(200).send('server reached');
-
+    email.sendCommentEmail(requestId, commenter, creator, gameTitle, system, function (err) {
+		if (err) {
+			res.status(500).send(err);
+		} else {
+			res.status(200).send('server reached');
+		}
+	});
 });
 
 router.get('/email/sendInviteRequest', function (req, res) {
